\name{B0}
\docType{data}
\alias{B0}
\title{Matrix indicating connections between the latent variables for QCAHS data.}
\description{Matrix indicating connections between the latent variables in the path model fitted to QCAHS data, used in the vignette to estimate the path coefficients of the model.
}
\usage{data(B0)}
\format{A square matrix of dimension 28.}
\keyword{Path matrix}