\name{GenPhen}
\docType{data}
\alias{GenPhen}
\title{Dataset to test GSCAestim and GSCA functions.}
\description{ Simulated data (for 999 individuals) of 4 SNPs mapped to 2 different genes and 4 traits involved in 2 different clinical pathways. The data is simulated such that one of the traits is involved in both clinical pathways and that one gene is connected to one of the clinical pathways and the other to both of them. See Figure 2, scenario (g), in Romdhani et al. (2014) for details.
}
\usage{data(GenPhen)}
\format{A data frame of 8 columns and 999 rows.}
\references{Romdhani, H., Hwang, H., Paradis, G., Roy-Gagnon, M.-H. and Labbe, A. (2014). Pathway-based Association Study of Multiple Candidate Genes and Multiple Traits Using Structural Equation Models, submitted.}
\keyword{datasets}