\name{QCAHS}
\docType{data}
\alias{QCAHS}
\title{Dataset used in the vignette.}
\description{ Dataset contaning some variables of interest from the Quebec Child and Adolescent Health and Social Survey (QCAHS), observed on $1707$ French Canadian participants (860 boys and 847 girls). Detailed descriptions of the QCAHS design and methods can be found in Paradis et al. (2003). The dataset contains 8 traits (z-score transformation standardized for age and sex), 33 SNPs and 2 polymorphisms with more than two alleles.
}
\usage{data(GenPhen)}
\format{A data frame of 49 columns and 1707 rows.}
\references{
Romdhani, H., Hwang, H., Paradis, G., Roy-Gagnon, M.-H. and Labbe, A. (2014). Pathway-based Association Study of Multiple Candidate Genes and Multiple Traits Using Structural Equation Models, submitted.
 Paradis, G., Lambert, M., O'Loughlin, J., Lavallee, C., Aubin, J., Berthiaume, P., Ledoux, M., Delvin, E., Levy, E., and Hanley, J. (2003). The quebec child and adolescent health and social survey: design and methods of a cardiovascular risk factor survey for youth. Can J Cardiol, 19:523-531.
}
\keyword{datasets}