\name{ResQCAHS}
\docType{data}
\alias{ResQCAHS}
\title{A list containing the results obtained for the QCAHS dataset.}
\description{ A list of 3 matrices: Weight containing the weight estimates, Path containing path coefficients estimates and pvalues containing p-values for all path coefficients.
}
\usage{data(ResQCAHS)}
\format{A list of 3 matrices.}
\keyword{Real data results.}