\name{W0}
\docType{data}
\alias{W0}
\title{Matrix indicating connections between the observed variables and the latent variables for QCAHS data.}
\description{Matrix indicating connections between the observed variables and the latent variables in the path model fitted to QCAHS data, used in the vignette to estimate the weight coefficients of the model.
}
\usage{data(W0)}
\format{A matrix of 49 columns and 28 rows.}
\keyword{Weight matrix}